+++ containers/display/content/quickCollectionPlotsAndFoldersHandling/get_filter_folders_and_plots.tsx
