+++ containers/display/content/quickCollectionPlotsAndFoldersHandling/get_filtered_folders_and_plots.tsx
