+++ containers/display/content/quickCollectionPlotsAndFoldersHandling/get_layout_from_quick_collections.tsx
