+++ containers/display/content/quickCollectionPlotsAndFoldersHandling/get_plots_friom_quick_collection.tsx
