+++ containers/display/content/quickCollectionPlotsAndFoldersHandling/use_get_folders_and_plots.tsx
